他的采购

+++ YHB_Prj/YHB_Prj/3rdChat/ChatView/MessageView/ChatCell/ChatCellBubble/EMChatAudioBubbleView.m
@@ -100,7 +100,7 @@ NSString *const kRouterEventAudioBubbleTapEventName = @"kRouterEventAudioBubbleT
 {
     [super setModel:model];
     
-    _timeLabel.text = [NSString stringWithFormat:@"%d'",self.model.time];
+    _timeLabel.text = [NSString stringWithFormat:@"%d'",(int)self.model.time];
     
     if (self.model.isSender) {
         [_isReadView setHidden:YES];
